Move MythicAPI class into separate file
This makes the code easier to call from elsewhere, e.g. if we want
to add additional scripts at a later date.

Also fixed a bug whereby the MythicAPI class was referring to the
args global variable, instead of the parameters passed in when
creating an instance (this didn't matter until the class was moved
into a separate file).
